Start working on an implementation of Karatsuba's Algorithm.
The exact multiplication of uinteger<N>'s should be implemented using
Karatsuba's Algorithm.

This commit is *NOT* already implementing it. Instead, the base case
(using uint64_t for uinteger<N> with N<=32) is added. In the wake of
thinking how to create uinteger<N> of appropriate size for the partial
products that arise in the algorithm (the prepend_zero_word method has
been implemented to test how C++ handles dependend types), it became
obvious that masking is not handled uniformly.


See #2