Fix bug where wrong operator== is called
operator== of word_array matches before operator== templated on a
template (probably because the former is a better match because for the
latter, the template template argument would need to be deduced). To resolve
this, I renamed the latter operator== to areIntegersEqual and provided
two separate operator==s that explicitly have integer<> resp. uinteger<>
arguments and both call areIntegersEqual.