Completely refactor the integer shift operators
The shift operators are now implemented in terms of the shift assignment
operators. This reduces code duplication.

Furthermore, the right shift is now directly implemented on word_arrays for
both, the logical and arithmetic right shift. The right shift assignment
operators of the signed and unsigned integers pick the correct versions.