Implement addition for posits
This implementation of addition uses and intermediate binprod type
that represents the posit as a product x * 2^k.

The advantage of this is that it is easy to implement for arbitrary bit
widths. Which is probably why other posit implementations use similar
approaches. The disadvantage is that it can be slow.

We will see if performance proves to be a problem. If it does, we can
use this slow implementation as a reference.