Introduce positparam and fraction classes
Positparam is a posit in the fraction * scale representation as used in
various literature.  Fraction is a simple fixed point fraction we use as
an aid when doing arithmetic on posits/positparam types.  Both of these
classes should eventually be hidden from the end user.

Notably, with this patch, addition for all positive values works
perfectly for all posit<8, 2> types.