Fix rounding for posit<8, 0> addition and subtraction.
By truncating less of the fraction, we can make better decisions about
when to round and when not to round. While this does work, it does
introduce the need for quite big scratch areas when computing the sums
of fractions.

A better approach would be to track a "truncated" bit. We should
investigate that approach.