Use constructors for creating posits and operators for converting.
This patch mostly just introduces a more easy to use API for
working with posits. In particular,

  auto p = posit16(x);

will now actually construct a posit that approximates the value of x
rather than a posit with bit pattern x. To import a bit pattern,
one should use the

  auto p = posit16::from(x);

factory methods provided by the posit class instead. Furthermore,
to convert a posit back to a standard type, one can now simply
use the overloaded conversion operators, viz.

  int n = int(p);
  double y = double(p);

and so on. Each of this conversions is explicit and assumes that the
user knows what they are doing. No bound checking is performed, so the
returned value might exhibit over- or underflow.