Port fixed_point code to work w/ clang.
The old code compiles just fine with GCC, but not with Clang.
Apparently there is some uncertainty whether the old code was valid C++
or not [1].  As even the current development build of Clang will not
compile the old version, we will probably have to stick with this
slightly uglier version.

[1] https://stackoverflow.com/questions/33872039/