Made CMake default to Release build
Be Advised:
This option is not ideal for debugging purposes. Either use
-DCMAKE_BUILD_TYPE=DEBUG or use cmake-gui to switch to a build with
debugging symbols.

Reasons:
This addition is to facilitate more accurate benchmarks by
preventing the Google benchmark library to be built with debug symbols.

Side effect:
When a new user builds the library no debug information is included by
default. This may give a better impression of the performance.
Also Projects that include this library need to worry less about options
and installation is simplified.